requirement.txt
requests
pixivpy
